refactor:"레디스 관련 설정 추가,변경"
