+++ src/main/resources/application.properties
@@ -19,7 +19,10 @@ cloud.aws.stack.auto=false
 jwt.secret.key=${jwt.secret.key}
 
 
-# app_password? ??? ????
+spring.data.redis.host=localhost
+spring.data.redis.port=6379
+
+
 spring.mail.host=smtp.gmail.com
 spring.mail.port= 587
 spring.mail.username=${gmail}
@@ -30,6 +33,5 @@ spring.mail.properties.mail.smtp.starttls.required=true
 spring.mail.properties.mail.smtp.connectiontimeout=5000
 spring.mail.properties.mail.smtp.timeout=5000
 spring.mail.properties.mail.smtp.writetimeout=5000
-# ? ??? ?? ??? ?? ?? ??? ????
 #spring.auth-code-expiration-millis=600,000
